Move time travel machine code to a dedicated module
