+++ timetravelmachine/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.github.vase4kin.timetravelmachine" />
