+++ timetravelmachine/src/test/java/com/github/vase4kin/timetravelmachine/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package com.github.vase4kin.timetravelmachine
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see [Testing documentation](http://d.android.com/tools/testing)
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, (2 + 2).toLong())
+    }
+}
